REFACTOR: move draft logic its own service
I took some inspiration for the syntax from https://github.com/discourse/discourse/blob/a8163a5c0cde06220ad71b2f96a304c28631a38a/app/assets/javascripts/discourse/app/services/presence.js#L246
Tried replacing init with constructor but the injected stuff can't be accessed i.e. this.currentUser
