+++ assets/javascripts/discourse/services/chat.js
@@ -63,13 +63,6 @@ export default Service.extend({
       this._subscribeToChannelEdits();
       this._subscribeToChannelStatusChange();
       this.presenceChannel = this.presence.getChannel("/chat/online");
-      this.draftStore = {};
-
-      if (this.currentUser.chat_drafts) {
-        this.currentUser.chat_drafts.forEach((draft) => {
-          this.draftStore[draft.channel_id] = JSON.parse(draft.data);
-        });
-      }
     }
   },
 
@@ -831,39 +824,6 @@ export default Service.extend({
 
   getDmChannelForUsernames(usernames) {
     return ajax("/chat/direct_messages.json", { data: { usernames } });
-  },
-
-  _saveDraft(channelId, draft) {
-    const data = { channel_id: channelId };
-    if (draft) {
-      data.data = JSON.stringify(draft);
-    }
-
-    ajax("/chat/drafts", { type: "POST", data });
-  },
-
-  setDraftForChannel(channel, draft) {
-    if (
-      draft &&
-      (draft.value || draft.uploads.length > 0 || draft.replyToMsg)
-    ) {
-      this.draftStore[channel.id] = draft;
-    } else {
-      delete this.draftStore[channel.id];
-      draft = null; // _saveDraft will destroy draft
-    }
-
-    discourseDebounce(this, this._saveDraft, channel.id, draft, 2000);
-  },
-
-  getDraftForChannel(channelId) {
-    return (
-      this.draftStore[channelId] || {
-        value: "",
-        uploads: [],
-        replyToMsg: null,
-      }
-    );
   },
 
   addToolbarButton() {
